backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.0
qdrant-client==1.10.0
google-generativeai==0.8.3
httpx==0.27.0
redis==5.0.0
gitpython==3.1.43
python-multipart==0.0.9
unstructured==0.15.0
tree-sitter==0.21.3
pydantic==2.8.0
